requirements.txt
mysql-connector-python==8.0.26
python-dotenv==0.19.0
Wikipedia-API==0.5.4
